Add insecure (http) mode to OpenWeather
I was working on a simple project and tried to get the weather. The problem was, this caused a stack overflow error, because of a core flaw in the BearSSL (my best understanding is that it times out and fills the stack, the reason it crashed is that I had little stack available) So I decided to modify the library to pass another parameter in the function called secure. When it is set to false, the client is not declared with BearSSL, but just a standard WIFIClient. When it is true, there is no modification to the original code.

Secure GET
ow.getForecast(current, hourly, daily, api_key, latitude, longitude, units, language, true);

Insecure and lighter GET
ow.getForecast(current, hourly, daily, api_key, latitude, longitude, units, language, false)

New parameter:
ow.getForecast(current, hourly, daily, api_key, latitude, longitude, units, language, -->bool<--)